Fix time bugs.
In practice, the recent time code I've added seems to work very well, but
this is probably just luck because it should be really broken. Let's fix it.

+++ athame.c
@@ -223,7 +223,7 @@ char athame_loop(int instream)
     else {
       while(selected == 0 && !athame_failure)
       {
-        int timeout_msec = get_timeout_msec();
+        long timeout_msec = get_timeout_msec();
         selected = athame_select(instream, vim_term, 0, timeout_msec, 0);
         if (waitpid(vim_pid, NULL, WNOHANG) == 0) // Is vim still running?
         {
